refactor: 컴포넌트 별 폴더 구조화
