+++ src/components/Card/Card.jsx
@@ -1,4 +1,4 @@
-import "../components/Card.css";
+import "../Card/Card.css";
 
 function Card() {
   return <div className="card-container"></div>;
