+++ src/components/User.jsx
@@ -11,6 +11,7 @@ function User() {
           <img
             className="user-avatar"
             src={dataFolder.folder.owner.profileImageSource}
+            alt="유저 이미지"
           />
           <p className="user-oner">@{dataFolder.folder.owner.name}</p>
           <p className="user-name">{dataFolder.folder.name}</p>
