+++ src/util/axiosInstance.js
@@ -1,0 +1,5 @@
+import axios from "axios";
+
+export const axiosInstance = axios.create({
+  baseURL: "https://bootcamp-api.codeit.kr/api/",
+});
